确认订单列表

+++ src/manage/address/index.js
@@ -1,5 +1,5 @@
-import React, { Component } from "react";
-import { Button, Icon, Table, message, Popconfirm, Alert } from "antd";
+import React, {Component} from "react";
+import {Button, Icon, Table, message, Popconfirm, Alert, Tooltip} from "antd";
 import FetchUtil from "../../utils/FetchUtil";
 import AddressForm from "./addressForm";
 import './address.css'
@@ -8,21 +8,31 @@ const ButtonGroup = Button.Group;
 export default class Address extends Component {
     columns = [
         {
-            title: '姓名',
+            title: 'お名前',
             dataIndex: 'name',
             key: 'name',
+            render: (name, {id, isDefault}, index) => {
+                return <span>
+                    {isDefault ?
+                        <Tooltip title="デフォルト"><Icon type="heart" theme="filled"
+                                                     style={{color: '#eb2f96'}}/> </Tooltip> :
+                        <Tooltip title="デフォルトにする"><Icon type="heart"
+                                                        onClick={() => this.setDefault(id, index)}/> </Tooltip>}
+                    {name}
+                </span>;
+            }
         },
         {
-            title: '地址',
+            title: 'アドレス',
             dataIndex: 'detailedAddress',
             key: 'detailedAddress',
             render: (detailedAddress, record) => {
-                const { province, city, district } = record;
+                const {province, city, district} = record;
                 return `${province} ${city} ${district} ${detailedAddress}`;
             }
         },
         {
-            title: '电话',
+            title: '電話番号',
             dataIndex: 'phone',
             key: 'phone',
         },
@@ -33,15 +43,20 @@ export default class Address extends Component {
             align: 'center',
             render: (id, record, index) => {
                 return <ButtonGroup>
-                    <Button title="修改" onClick={() => this.openEdit(record, index)}><Icon type="edit" /></Button>
-                    <Popconfirm
-                        title="确定删除吗？"
-                        onConfirm={() => this.removeAddress(id, index)}
-                        okText="是"
-                        cancelText="否"
-                    >
-                        <Button title="删除" type="danger"><Icon type="delete" /></Button>
-                    </Popconfirm>
+                    <Button title="修正" onClick={() => this.openEdit(record, index)}><Icon type="edit"/></Button>
+                    {
+                        record.isDefault ?
+                            <Button title="削除" type="danger" disabled={true}><Icon type="delete"/></Button>
+                            :
+                            <Popconfirm
+                                title="削除しますか？"
+                                onConfirm={() => this.removeAddress(id, index)}
+                                okText="はい"
+                                cancelText="いいえ"
+                            >
+                                <Button title="削除" type="danger"><Icon type="delete"/></Button>
+                            </Popconfirm>
+                    }
                 </ButtonGroup>
             }
         }
@@ -52,67 +67,98 @@ export default class Address extends Component {
         address: {},
         maxCount: 0
     }
+
     findAddressList() {
         FetchUtil.get({
             url: '/user/address/list',
-            success: ({ data: addressList }) => {
-                this.setState({ addressList });
+            success: ({data: addressList}) => {
+                this.setState({addressList});
             }
         })
     }
+
     componentWillMount() {
         this.getMaxCount();
         this.findAddressList();
     }
+
     openAdd() {
-        this.setState({ formVisible: true, address: {} })
+        this.setState({formVisible: true, address: {}})
     }
+
     openEdit(address, editIndex) {
-        this.setState({ formVisible: true, address, editIndex })
+        this.setState({formVisible: true, address, editIndex})
     }
+
     saveAddress(address) {
         FetchUtil.put({
             url: '/user/address/save',
             data: address,
-            success: ({ data }) => {
-                const { addressList, editIndex } = this.state;
+            success: ({data}) => {
+                message.info('セーブ完了^_^');
+                const {addressList, editIndex} = this.state;
                 if (!address.id) {
                     address.id = data;
+                    address.isDefault = addressList.length === 0 ? 1 : 0;
                     addressList.push(address);
                 } else {
                     addressList[editIndex] = address;
                 }
-                this.setState({ addressList, formVisible: false });
-                message.info('保存成功啦^_^');
+                this.setState({addressList, formVisible: false});
             }
         })
     }
+
     removeAddress(id, index) {
+        const {addressList} = this.state;
+        if (addressList[index].isDefault) {
+            message.warn('デフォルトの配送先を削除するのはダメだよ')
+            return;
+        }
         FetchUtil.delete({
             url: `/user/address/${id}`,
             success: () => {
-                const { addressList } = this.state;
                 addressList.splice(index, 1);
-                this.setState({ addressList });
-                message.info('地址已经删除啦^_^');
+                this.setState({addressList});
+                message.info('削除完了^_^');
             }
         })
     }
+
     getMaxCount() {
         FetchUtil.get({
             url: '/basic/address/maxCount',
-            success: ({ data: maxCount }) => this.setState({ maxCount })
+            success: ({data: maxCount}) => this.setState({maxCount})
         });
     }
+
+    setDefault(id, curIndex) {
+        FetchUtil.put({
+            url: `/user/address/${id}/default`,
+            success: () => {
+                const {addressList} = this.state;
+                addressList.map((record, index) => {
+                    if (record.isDefault) {
+                        record.isDefault = 0;
+                    } else if (curIndex === index) {
+                        record.isDefault = 1;
+                    }
+                });
+                this.setState({addressList});
+            }
+        })
+
+    }
+
     render() {
-        const { formVisible, addressList, address, maxCount } = this.state;
+        const {formVisible, addressList, address, maxCount} = this.state;
         return <main>
             <Alert
-                style={{ marginBottom: '10px' }}
+                style={{marginBottom: '10px'}}
                 type="info"
-                message={`共 ${addressList.length} 条，最多 ${maxCount} 条地址`} />
+                message={`合計 ${addressList.length} アドレス，最大 ${maxCount} アドレス`}/>
             <Button shape="circle" className="addBtn" onClick={() => this.openAdd()}>
-                <Icon type="plus" />
+                <Icon type="plus"/>
             </Button>
             <Table
                 pagination={false}
@@ -123,7 +169,7 @@ export default class Address extends Component {
             <AddressForm
                 visible={formVisible}
                 address={address}
-                submit={(newAddress) => this.saveAddress(newAddress)} />
+                submit={(newAddress) => this.saveAddress(newAddress)}/>
         </main>
     }
 }
